ArmedGrenade is part of Grenade
