+++ src/items.rs
@@ -1,7 +1,6 @@
 //! Things available to spawn from the level editor
 //! Proto-mods, eventually some of the items will move to some sort of a wasm runtime
 
-mod armed_grenade;
 mod grenades;
 mod machine_gun;
 mod muscet;
