+++ src/items/grenades.rs
@@ -1,10 +1,11 @@
 use macroquad::{
+    color,
     //audio::play_sound_once,
     experimental::{
         animation::{AnimatedSprite, Animation},
         collections::storage,
         coroutines::{start_coroutine, wait_seconds, Coroutine},
-        scene::{self, Handle, HandleUntyped, RefMut},
+        scene::{self, Handle, HandleUntyped, Node, RefMut},
     },
     prelude::*,
 };
@@ -12,7 +13,6 @@ use macroquad::{
 use crate::{
     capabilities,
     components::{GunlikeAnimation, PhysicsBody, ThrowableItem},
-    items::armed_grenade::ArmedGrenade,
     nodes::Player,
     Resources,
 };
@@ -248,5 +248,122 @@ impl scene::Node for Grenades {
         if !node.throwable.thrown() {
             node.draw_hud();
         }
+    }
+}
+
+struct ArmedGrenade {
+    grenade_sprite: AnimatedSprite,
+    pub body: PhysicsBody,
+    lived: f32,
+}
+
+impl ArmedGrenade {
+    pub const COUNTDOWN_DURATION: f32 = 0.5;
+    pub const EXPLOSION_RADIUS: f32 = 100.0;
+
+    pub fn new(pos: Vec2, facing: bool) -> Self {
+        // TODO: In case we want to animate thrown grenades rotating etc.
+        let grenade_sprite = AnimatedSprite::new(
+            21,
+            28,
+            &[Animation {
+                name: "idle".to_string(),
+                row: 1,
+                frames: 4,
+                fps: 8,
+            }],
+            false,
+        );
+
+        let speed = if facing {
+            vec2(600., -200.)
+        } else {
+            vec2(-600., -200.)
+        };
+
+        let mut resources = storage::get_mut::<Resources>();
+
+        let grenade_mount_pos = if facing {
+            vec2(30., 10.)
+        } else {
+            vec2(-50., 10.)
+        };
+
+        let size = vec2(15.0, 15.0);
+
+        let body = PhysicsBody {
+            pos,
+            facing,
+            angle: 0.0,
+            size,
+            speed,
+            collider: resources.collision_world.add_actor(
+                pos + grenade_mount_pos,
+                size.x as _,
+                size.y as _,
+            ),
+            on_ground: false,
+            last_frame_on_ground: false,
+            have_gravity: true,
+            bouncyness: 0.5,
+        };
+
+        ArmedGrenade {
+            grenade_sprite,
+            body,
+            lived: 0.0,
+        }
+    }
+
+    pub fn spawn(pos: Vec2, facing: bool) {
+        let grenade = ArmedGrenade::new(pos, facing);
+        scene::add_node(grenade);
+    }
+}
+
+impl Node for ArmedGrenade {
+    fn fixed_update(mut node: RefMut<Self>) {
+        node.grenade_sprite.update();
+
+        node.body.update();
+
+        node.lived += get_frame_time();
+
+        if node.lived >= ArmedGrenade::COUNTDOWN_DURATION {
+            {
+                let mut resources = storage::get_mut::<Resources>();
+                resources.hit_fxses.spawn(node.body.pos);
+            }
+            let grenade_circ = Circle::new(
+                node.body.pos.x,
+                node.body.pos.y,
+                ArmedGrenade::EXPLOSION_RADIUS,
+            );
+            for mut player in scene::find_nodes_by_type::<crate::nodes::Player>() {
+                if grenade_circ.overlaps_rect(&player.get_hitbox()) {
+                    let direction = node.body.pos.x > (player.body.pos.x + 10.);
+                    scene::find_node_by_type::<crate::nodes::Camera>().unwrap();
+                    player.kill(direction);
+                }
+            }
+            node.delete();
+        }
+    }
+
+    fn draw(node: RefMut<Self>) {
+        let resources = storage::get_mut::<Resources>();
+        draw_texture_ex(
+            resources.items_textures["grenades/explosives"],
+            node.body.pos.x,
+            node.body.pos.y,
+            color::WHITE,
+            DrawTextureParams {
+                source: Some(node.grenade_sprite.frame().source_rect),
+                dest_size: Some(node.grenade_sprite.frame().dest_size),
+                flip_x: false,
+                rotation: 0.0,
+                ..Default::default()
+            },
+        );
     }
 }
